Grouped bars support.
This is already supported, in theory.  We can draw bars and if the data
shows the bars as grouped, they will be grouped.  This implementation
makes the bars chart type aware of series, and in particular other
series when drawing.  I have not implemented hit suppoort (this presents
a similar coupling problem).

Right now, I'm leaning towards doing this with data preprocessing.  In
the current architecture, that preprocessing could be done with a plugin.
However, the drawback here is performance.  I'll think on this one more.
Let me know if you have any ideas.

Fixes #65.