Add package.json for dependencies
This allows contributors to easily install development dependencies via
npm install. The 'private' property prevents the project from being
listed as a package in the npm registry (as it is not actually
compatible with Node just yet).

Also added npm-debug.log to the project .gitignore so it does not get
tracked since it's just a by-product of npm.

This will probably also need jasmine and other dependencies as this is
just a foundation.

Also, note that since flotr2 is not yet Node-compatible, it should not
be installed or used as an npm package. We are simply leveraging the
dependency management features.