Add smoosh dependencies to Makefile
Setting the SMOOSH variable to the path of the binary allows us a
shorthand reference for use in a rule, listing as a dependency, and even
defining a target.

I've added a target so that if smoosh isn't there, make will actually
trigger NPM.

Then every target that used smoosh was changed over to use the $(SMOOSH)
variable (No more $PATH problem!) and given smoosh as a dependency.